Use proper revision option when reading stream from start or end
When from offset is set to 0, first event in the stream is skipped as 0
is considered the last seen event. Instead, we need to use the proper
`api.ReadReq_Options_StreamOptions_Start` type to consume events from
the start of the stream.

When from offset is set to max uint64 (i.e. `^uint64(0)`), we need to
use the proper `api.ReadReq_Options_StreamOptions_End` type to consume
events from the end of the stream.